Manage weapons
Even with all the refactoring I did, there is still a ton of overlap
between character and weapon copies. As much as I want to go ahead with
another refactor, it is probably best that I don't for now since they
may behave a lot differently as I pack on new features.